jsonrpc: move qr/uri type to dedicated file
#skip-changelog
